+++ packages/block-editor/src/components/block-list/index.js
@@ -7,7 +7,7 @@ import classnames from 'classnames';
  * WordPress dependencies
  */
 import { useRef } from '@wordpress/element';
-import { AsyncModeProvider, useSelect } from '@wordpress/data';
+import { AsyncModeProvider, useSelect, useDispatch } from '@wordpress/data';
 
 /**
  * Internal dependencies
@@ -16,6 +16,9 @@ import BlockListBlock from './block';
 import BlockListAppender from '../block-list-appender';
 import __experimentalBlockListFooter from '../block-list-footer';
 import useMultiSelection from './use-multi-selection';
+import { getBlockClientId } from '../../utils/dom';
+
+/** @typedef {import('@wordpress/element').WPSyntheticEvent} WPSyntheticEvent */
 
 /**
  * If the block count exceeds the threshold, we disable the reordering animation
@@ -71,6 +74,7 @@ function BlockList( {
 		hasMultiSelection,
 		enableAnimation,
 	} = useSelect( selector, [ rootClientId ] );
+	const { selectBlock } = useDispatch( 'core/block-editor' );
 	const ref = useRef();
 	const onSelectionStart = useMultiSelection( { ref, rootClientId } );
 
@@ -80,6 +84,45 @@ function BlockList( {
 		...__experimentalUIParts,
 	};
 
+	let onFocus;
+	let onDragStart;
+
+	if ( ! rootClientId ) {
+		/**
+		 * Marks the block as selected when focused and not already selected. This
+		 * specifically handles the case where block does not set focus on its own
+		 * (via `setFocus`), typically if there is no focusable input in the block.
+		 *
+		 * @param {WPSyntheticEvent} event
+		 */
+		onFocus = ( event ) => {
+			if ( hasMultiSelection ) {
+				return;
+			}
+
+			const clientId = getBlockClientId( event.target );
+
+			if ( clientId && clientId !== selectedBlockClientId ) {
+				selectBlock( clientId );
+			}
+		};
+
+		/**
+		 * Prevents default dragging behavior within a block.
+		 * To do: we must handle this in the future and clean up the drag target.
+		 * Previously dragging was prevented for multi-selected, but this is no longer
+		 * needed.
+		 *
+		 * @param {WPSyntheticEvent} event Synthetic drag event.
+		 */
+		onDragStart = ( event ) => {
+			// Ensure we target block content, not block controls.
+			if ( getBlockClientId( event.target ) ) {
+				event.preventDefault();
+			}
+		};
+	}
+
 	return (
 		<div
 			ref={ ref }
@@ -87,6 +130,8 @@ function BlockList( {
 				'block-editor-block-list__layout',
 				className
 			) }
+			onFocus={ onFocus }
+			onDragStart={ onDragStart }
 		>
 			{ blockClientIds.map( ( clientId, index ) => {
 				const isBlockInSelection = hasMultiSelection ?
